Make chargeRate in ChargeState to Double

+++ Source/Models/ChargeState.swift
@@ -21,7 +21,7 @@ public class ChargeState {
     public let chargeMilesAddedRated: Double
     public let chargePortDoorOpen: Bool
     public let chargePortLatch: String
-    public let chargeRate: Int
+    public let chargeRate: Double
     public let chargeToMaxRange: Int
     public let chargerActualCurrent: Int
     public let chargerPhases: Int?
@@ -64,7 +64,7 @@ public class ChargeState {
         chargeMilesAddedRated = dictionary["charge_miles_added_rated"] as! Double
         chargePortDoorOpen = dictionary["charge_port_door_open"] as! Bool
         chargePortLatch = dictionary["charge_port_latch"] as! String
-        chargeRate = dictionary["charge_rate"] as! Int
+        chargeRate = dictionary["charge_rate"] as! Double
         chargeToMaxRange = dictionary["charge_to_max_range"] as! Int
         chargerActualCurrent = dictionary["charger_actual_current"] as! Int
         chargerPhases = dictionary["charger_phases"] as? Int
